pyproject.toml
[tool.poetry]
name = "skymarket"
version = "0.1.0"
description = ""
authors = ["Deadpo_Ol <https://github.com/devoind>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
drf-spectacular = "^0.24.2"
djangorestframework-simplejwt = "^5.2.2"
psycopg2 = "^2.9.5"
django-phonenumber-field = "^7.0.0"
django-phonenumbers = "^1.0.1"
django-rest-framework-nested = "^0.0.1"
sentry-sdk = "^1.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
